requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
aiosqlite==0.21.0
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
apprise==1.9.2
asgi-lifespan==2.1.0
asgiref==3.8.1
asttokens==3.0.0
asyncpg==0.30.0
attrs==25.1.0
beautifulsoup4==4.12.3
cachetools==5.5.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.2
coolname==2.2.0
croniter==6.0.0
cryptography==44.0.1
dateparser==1.2.1
debugpy==1.8.12
decorator==5.1.1
Deprecated==1.2.18
Django==5.1.6
django-environ==0.12.0
django-storages==1.14.4
dnspython==2.7.0
docker==7.1.0
exceptiongroup==1.2.2
executing==2.2.0
Faker==36.1.1
fastapi==0.115.8
frozenlist==1.5.0
fsspec==2025.2.0
gcsfs==2025.2.0
google-api-core==2.24.1
google-auth==2.38.0
google-auth-oauthlib==1.2.1
google-cloud-core==2.4.1
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
graphviz==0.20.3
greenlet==3.1.1
griffe==1.5.7
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
humanize==4.11.0
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.5.0
ipykernel==6.29.5
ipython==8.31.0
jedi==0.19.2
Jinja2==3.1.5
jinja2-humanize-extension==0.4.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
Mako==1.3.9
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
more-itertools==10.6.0
multidict==6.1.0
nest-asyncio==1.6.0
numpy==2.2.2
oauthlib==3.2.2
opentelemetry-api==1.30.0
orjson==3.10.15
packaging==24.2
pandas==2.2.3
parso==0.8.4
pathspec==0.12.1
pendulum==3.0.0
pexpect==4.9.0
pillow==11.1.0
pip3-autoremove==1.2.2
platformdirs==4.3.6
playwright==1.49.1
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.2.1
proto-plus==1.26.0
protobuf==4.25.6
psutil==6.1.1
psycopg==3.2.4
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic-extra-types==2.10.2
pydantic-settings==2.7.1
pydantic_core==2.27.2
pyee==12.0.0
Pygments==2.19.1
pymongo==4.11.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
python-socks==2.7.1
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
readchar==4.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
setuptools==75.8.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.45.3
text-unidecode==1.3
time-machine==2.16.0
toml==0.10.2
tornado==6.4.2
traitlets==5.14.3
typer==0.15.1
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3
ujson==5.10.0
urllib3==2.3.0
uvicorn==0.34.0
wcwidth==0.2.13
websockets==13.1
wrapt==1.17.2
yarl==1.18.3
zipp==3.21.0
